refactor: rename folder
frontend -> web
