+++ web/src/components/sign-in-form.tsx
@@ -2,8 +2,8 @@ import { useForm } from '@tanstack/react-form'
 import { useNavigate } from 'react-router'
 import { toast } from 'sonner'
 import z from 'zod'
-import InputComponent from '../components/input.component'
 import { authClient } from '../lib/auth-client'
+import InputComponent from './input.component'
 import { Button } from './ui/button'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from './ui/card'
 import { Spinner } from './ui/spinner'
