+++ web/src/utils/calculations.ts
